+++ apps/ssr-isr/server.ts
@@ -63,7 +63,10 @@ export function app(): express.Express {
   server.get(
     '*',
     // Serve page if it exists in cache
-    async (req, res, next) => await isr.serveFromCache(req, res, next),
+    async (req, res, next) =>
+      await isr.serveFromCache(req, res, next, {
+        providers: [{ provide: RESPONSE, useValue: res }],
+      }),
 
     // Server side render the page and add to cache if needed
     async (req, res, next) =>